Split gengo.go and move func rename
